+++ src/dune/scope.ml
@@ -39,29 +39,42 @@ module DB = struct
   let find_by_project t project =
     Path.Source.Map.find_exn t.by_dir (Dune_project.root project)
 
+  type redirect_to =
+    | Project of Dune_project.t
+    | Name of Lib_name.t
+
   let resolve t public_libs name : Lib.DB.Resolve_result.t =
     match Lib_name.Map.find public_libs name with
     | None -> Not_found
-    | Some project ->
+    | Some (Project project) ->
       let scope = find_by_project (Fdecl.get t) project in
       Redirect (Some scope.db, name)
+    | Some (Name name) -> Redirect (None, name)
 
-  let public_libs t ~stdlib_dir ~installed_libs internal_libs =
+  let public_libs t ~stdlib_dir ~installed_libs stanzas =
     let public_libs =
-      List.filter_map internal_libs
-        ~f:(fun (_dir, (lib : Dune_file.Library.t)) ->
-          Option.map lib.public ~f:(fun p ->
-              (Dune_file.Public_lib.name p, lib.project)))
+      List.filter_map stanzas
+        ~f:(fun (stanza : Lib.DB.Library_related_stanza.t) ->
+          match stanza with
+          | Library (_, { project; public = Some p; _ }) ->
+            Some (Dune_file.Public_lib.name p, Project project)
+          | Library _ -> None
+          | External_variant _ -> None
+          | Deprecated_library_name x ->
+            Some
+              ( Dune_file.Public_lib.name x.old_public_name
+              , Name x.new_public_name ))
       |> Lib_name.Map.of_list
       |> function
       | Ok x -> x
       | Error (name, _, _) -> (
         match
-          List.filter_map internal_libs ~f:(fun (_dir, lib) ->
-              Option.bind lib.public ~f:(fun p ->
-                  Option.some_if
-                    (name = Dune_file.Public_lib.name p)
-                    lib.buildable.loc))
+          List.filter_map stanzas ~f:(fun stanza ->
+              match stanza with
+              | Library (_, { buildable = { loc; _ }; public = Some p; _ })
+               |Deprecated_library_name { loc; old_public_name = p; _ } ->
+                Option.some_if (name = Dune_file.Public_lib.name p) loc
+              | _ -> None)
         with
         | []
          |[ _ ] ->
@@ -78,39 +91,30 @@ module DB = struct
     Lib.DB.create () ~stdlib_dir ~parent:installed_libs ~resolve
       ~all:(fun () -> Lib_name.Map.keys public_libs)
 
-  let scopes_by_dir context ~projects ~lib_config ~public_libs internal_libs
-      variant_implementations =
+  let scopes_by_dir context ~projects ~lib_config ~public_libs stanzas =
     let build_context_dir = Path.Build.relative Path.Build.root context in
     let projects_by_dir =
       List.map projects ~f:(fun (project : Dune_project.t) ->
           (Dune_project.root project, project))
       |> Path.Source.Map.of_list_exn
     in
-    let libs_by_project_dir =
-      List.map internal_libs ~f:(fun (dir, (lib : Dune_file.Library.t)) ->
-          (Dune_project.root lib.project, (dir, lib)))
+    let stanzas_by_project_dir =
+      List.map stanzas ~f:(fun (stanza : Lib.DB.Library_related_stanza.t) ->
+          let project =
+            match stanza with
+            | Library (_, lib) -> lib.project
+            | External_variant ev -> ev.project
+            | Deprecated_library_name x -> x.project
+          in
+          (Dune_project.root project, stanza))
       |> Path.Source.Map.of_list_multi
     in
-    let variant_implementations_by_project_dir =
-      List.map variant_implementations
-        ~f:(fun (lib : Dune_file.External_variant.t) ->
-          (Dune_project.root lib.project, lib))
-      |> Path.Source.Map.of_list_multi
-    in
-    let libs_variants_by_project_dir =
-      Path.Source.Map.merge libs_by_project_dir
-        variant_implementations_by_project_dir ~f:(fun _dir libs variants ->
-          let libs = Option.value libs ~default:[] in
-          let variants = Option.value variants ~default:[] in
-          Some (libs, variants))
-    in
-    Path.Source.Map.merge projects_by_dir libs_variants_by_project_dir
-      ~f:(fun _dir project l_v ->
+    Path.Source.Map.merge projects_by_dir stanzas_by_project_dir
+      ~f:(fun _dir project stanzas ->
         let project = Option.value_exn project in
-        let libs, variants = Option.value l_v ~default:([], []) in
+        let stanzas = Option.value stanzas ~default:[] in
         let db =
-          Lib.DB.create_from_library_stanzas libs variants ~parent:public_libs
-            ~lib_config
+          Lib.DB.create_from_stanzas stanzas ~parent:public_libs ~lib_config
         in
         let root =
           Path.Build.append_source build_context_dir
@@ -118,16 +122,14 @@ module DB = struct
         in
         Some { project; db; root })
 
-  let create ~projects ~context ~installed_libs ~lib_config internal_libs
-      variant_implementations =
+  let create ~projects ~context ~installed_libs ~lib_config stanzas =
     let t = Fdecl.create () in
     let public_libs =
       public_libs t ~stdlib_dir:lib_config.Lib_config.stdlib_dir
-        ~installed_libs internal_libs
+        ~installed_libs stanzas
     in
     let by_dir =
-      scopes_by_dir context ~projects ~lib_config ~public_libs internal_libs
-        variant_implementations
+      scopes_by_dir context ~projects ~lib_config ~public_libs stanzas
     in
     let value = { by_dir; context } in
     Fdecl.set t value;